tests/wpt/web-platform-tests/presentation-api/controlling-ua/PresentationConnection_onterminated-manual.html

<!DOCTYPE html>

<meta charset="utf-8">
<title>PresentationConnection.onterminate</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="http://w3c.github.io/presentation-api/#starting-a-presentation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="common.js"></script>
<h2>Description</h2>
<p>
  This test validates that after connection terminate,<br/>
  the connection state is set terminated,<br/>
  the onterminate EventHandler is triggered.
</p>
<br/>
<p>Click the button below to start the test.</p>
<button id="presentBtn" onclick="startPresentation()">Start Presentation Test</button>

<script>
  setup({explicit_timeout: true});
  var startPresentation = function () {
    document.getElementById('presentBtn').disabled = true;
    async_test(function(t) {
      var request = new PresentationRequest(presentationUrls);
      request.start()
        .then(function(connection) {

          // Enable timeout again, cause no user action is needed from here.
          t.step_timeout(function() {
              t.force_timeout();
              t.done();
          }, 5000);

          assert_true(connection instanceof PresentationConnection);
          connection.onconnect = t.step_func(function(evt) {
            assert_equals(connection.state, "connected");
            connection.terminate();
          });
          connection.onterminate = t.step_func_done(function(evt) {
            assert_equals(evt.type, "terminate");
            assert_equals(connection.state, "terminated");
          });
          connection.onclose = t.step_func_done(function(evt) {
              assert_unreached("Wrong, the onclose shouldn't be triggered!");
          });
      })
      .catch(function(ex) {
        assert_unreached(ex.name + ":" + ex.message);
      });
    }, "the onterminate is fired and the connection state is terminated");
  }
</script>
